+++ tests/test_restore_models.py
@@ -3,7 +3,7 @@ import os
 
 import torch
 
-import tests.utils as tutils
+import tests.models.utils as tutils
 from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import ModelCheckpoint
 from tests.models import LightningTestModel
